pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "netintent-ops"
version = "0.1.0"
description = "Network intent operations platform"
readme = "README.md"
requires-python = ">=3.11"
authors = [
  {name = "NetIntent Ops Authors"},
]
license = {text = "Apache-2.0"}
dependencies = [
  "fastapi>=0.104.0",
  "uvicorn[standard]>=0.23.0",
  "pydantic>=2.4.2",
  "ansible-runner>=2.3.4",
  "hvac>=1.2.1",
  "sqlalchemy>=2.0.20",
  "psycopg2-binary>=2.9.9",
  "jinja2>=3.1.2",
  "pyyaml>=6.0.1",
  "aiofiles>=23.2.1",
  "python-dotenv>=1.0.0",
  "tenacity>=8.2.3",
  "rich>=13.6.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4.2",
  "molecule>=6.0.2",
  "molecule-plugins[docker,delegated]>=25.8.0",
  "pre-commit>=3.4.0"
]

[tool.setuptools]
packages = ["api", "worker"]

[tool.setuptools.package-data]
"ansible" = ["**/*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
